requirements.txt
python-telegram-bot==13.15
requests==2.28.2
python-dotenv==1.0.0
cryptography==39.0.1
